requirements.txt
discord.py==2.6.3
gspread==6.2.1
oauth2client==4.1.2